Report our run everywhere: timber, water use and flood tasks compute their numbers
The committed tables move to comparison anchors (logged in each task, pinned in tests).
Timber sums the verified value raster on the country-id raster (88,743,966,355 vs the
committed 88,743,967,091). Water use reports the chain's own components (agriculture
1.065e12 over 126 countries, all-sector 3.661e13 over 154; 20 chain countries pending
their r250 name match). Flood computes per-country EAD through the ported chain (the
run regenerating the table is detached). Suite 149 green.

+++ global_invest/timber_provision/timber_provision_chain.py
@@ -88,5 +88,5 @@ def timber_gep_from_zone_sums(zone_sums, countries_df):
         pd.DataFrame: countries_df plus a timber_provision_gep column.
     """
     df = countries_df.copy()
-    df["timber_provision_gep"] = zone_sums[df["iso3_r250_id"].to_numpy()]
+    df["timber_provision_gep"] = zone_sums[df["iso3_r250_id"].to_numpy(dtype=int)]
     return df
